requirements.txt

Django==3.0.5
djangorestframework==3.11.0
mysqlclient==1.4.6
psycopg2==2.8.5

